Extract logic of observing the selected range to an external usecase
